idadMain/requirements.txt
torch>=1.8.0
torchvision
torchsde
pyro-ppl==1.6.0
mlflow
pandas
numpy
